+++ src/main/java/com/prj1/controller/MemberController.java
@@ -3,6 +3,8 @@ package com.prj1.controller;
 import com.prj1.domain.Member;
 import com.prj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,38 +17,46 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 @RequiredArgsConstructor
 @RequestMapping("member")
 public class MemberController {
+
     private final MemberService service;
 
     @GetMapping("signup")
     public String signupForm() {
-        // 리턴하는게 같으니까 void로 해도 됨
+
         return "member/signup";
     }
 
     @PostMapping("signup")
     public String signup(Member member) {
         service.signup(member);
-        return "redirect:/";
 
+        return "redirect:/";
     }
 
     @GetMapping("list")
+    @PreAuthorize("hasAuthority('admin')")
     public String list(Model model) {
         model.addAttribute("memberList", service.list());
         return "member/list";
     }
 
     @GetMapping("")
-    public String info(Integer id, Model model) {
-        model.addAttribute("member", service.get(id));
-        return "member/info";
+    public String info(Integer id, Authentication authentication, Model model) {
+        if (service.hasAccess(id, authentication)
+                || service.isAdmin(authentication)) {
+            model.addAttribute("member", service.get(id));
+            return "member/info";
+        }
+        return "redirect:/";
     }
 
     @PostMapping("remove")
-    public String remove(Integer id) {
-        service.remove(id);
+    public String remove(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
 
-        return "redirect:/member/signup";
+        return "redirect:/logout";
     }
 
     @GetMapping("modify")
@@ -57,8 +67,10 @@ public class MemberController {
     }
 
     @PostMapping("modify")
-    public String modify(Member member, RedirectAttributes rttr) {
-        service.modify(member);
+    public String modify(Member member, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(member.getId(), authentication)) {
+            service.modify(member);
+        }
 
         rttr.addAttribute("id", member.getId());
         return "redirect:/member";
@@ -71,5 +83,9 @@ public class MemberController {
         String message = service.emailCheck(email);
         return message;
     }
-}
 
+    @GetMapping("login")
+    public String loginForm() {
+        return "member/login";
+    }
+}
